refactor: promote AboutPageLayout to app layer's ui
