+++ apps/billets-web/app/(ui)/about-page-layout/mobile-menu/mobile-menu.types.ts
