+++ apps/billets-web/app/about/(ui)/index.ts
@@ -4,6 +4,3 @@ export * from './landing-image';
 export * from './landing-section';
 export * from './landing-text';
 export * from './landing-youtube';
-export * from './header';
-export * from './mobile-menu';
-export * from './about-page-layout';
