+++ apps/billets-web/app/store/page.tsx
@@ -1,4 +1,4 @@
-import { AboutPageLayout } from 'app/about/(ui)';
+import { AboutPageLayout } from 'app/(ui)/about-page-layout';
 
 export default function StorePage() {
   return <AboutPageLayout>StorePage</AboutPageLayout>;
